tasks.py: Fix env var not updating when used also in a dependsOn task
If the same env var appeared in a task and in a dependsOn task of this
task, like the VSCODE_CMD on tcb-platform-push-ostree, with the
tcb-unpack as dependsOn, the env var was not getting updated
correctly when using the TASKS_OVERRIDE_ENV: "True" var (which is used
on CI/CD). This commit fixes it.

Signed-off-by: Andre Riesco <andre.riesco@toradex.com>