templates: Improve how tasks are merged
The tcb template had a unique behavior to merge its tasks.json with common.json. This behavior required the tasks to be duplicated in the tasks.json and was inefficient. By modifying the merge logic, it is now possible to merge tasks in a granular way that is more efficient and doesn't require specific conditions for the tcb template. This will also be more efficient for future templates that may not need to merge all tasks from common.json.

Related-to: TIE-915
Signed-off-by: Leonardo Bornia <leonardo.bornia@toradex.com>