scripts: validate-deps-running:  Add a fallback to enable binfmt
There are some really special cases where the binfmt is not enabled,
even though the initialization of the template should have enabled it.
This commit adds a fallback to enable the binfmt if it is not enabled
after the initialization of the template.
This is done by running the binfmt container again, which should
enable the binfmt. This is done in a retry loop, so if the first
attempt fails, it will try again. This is done to avoid the user
having to run the binfmt container manually.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>