tasks: Drop usage of the registry to host -> remote exchange
Instead let's use a local store folder to exchange the container image
with the remote host. Using the docker save/load commands combined with
rsync to perform optimal transfer of the image.

Another gain point is the ability to not use compressed layers, which
can be useful when using cloud registries, but for local development
it is not needed and can save some processing time under our more
less powerful machines.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>