rustSlint: cargo: Disable debug symbols in dev profile
This will reduce the binary size considerably for debug builds. As
the developer at first need the debug symbols to be present only
for their own code, not for dependencies.

Also this will speed up remote sharing and deployment, as the
binaries will be smaller, improving the remote development
experience.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>