tasks/launch: Adjust flow to fix race condition for spin-down-tunnel
spin-down-tunnel task can trigger race conditions on multi-root
workspaces or when multiple debug sessions are started in parallel.
This patch moves the spin-down-tunnel task to be a postDebugTask
of each launch configuration and also adds it as the last task
in the sequences that build and deploy containers to ensure that
the tunnel is always spun down after all the operations are done.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>