tasks/common/scripts: Add spin-up-down-registry xonsh script
For multi-container projects the spin down of the torizon-ide-port-tunnel
container could lead to race conditions where one task spins down the
container while another task is trying to spin it up or using it.
To avoid this we introduce a locker mechanism in a new script
spin-up-down-registry.xsh that will keep track of how many tasks
are using the container and only spin it down when no task is using it.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>